replace some more ulong with uintptr

+++ utils/c2l/lex.c
@@ -1473,7 +1473,7 @@ alloc(long n)
 {
 	void *p;
 
-	while((ulong)hunk & MAXALIGN) {
+	while((uintptr)hunk & MAXALIGN) {
 		hunk++;
 		nhunk--;
 	}
